refactor: 여러 군데에서 사용하는 상수를 Constant 클래스로 분리
