+++ src/main/kotlin/lotto/model/Lotto.kt
@@ -1,10 +1,11 @@
 package lotto.model
 
-import org.assertj.core.util.Lists
+import lotto.util.Constants.LOTTO_SIZE
+import lotto.util.Constants.MATCH_FIVE_BONUS
 
 class Lotto(private val numbers: List<Int>) {
     init {
-        require(numbers.size == 6)
+        require(numbers.size == LOTTO_SIZE)
     }
 
     fun checkLottoWinning(bonusNum:Int,lottoLists: List<List<Int>>):List<Int>{
@@ -14,7 +15,7 @@ class Lotto(private val numbers: List<Int>) {
            val a=  lottoLists[i].intersect(numbers.toSet()).count()
             result[i] = a
             if(result[i] == 5 && lottoLists[i].contains(bonusNum)){
-                result[i] = 100
+                result[i] = MATCH_FIVE_BONUS
             }
         }
         return result
