+++ src/main/kotlin/lotto/model/LottoResultAnalyzer.kt
@@ -1,16 +1,16 @@
 package lotto.model
 
+import lotto.util.Constants.MATCH_FIVE
+import lotto.util.Constants.MATCH_FIVE_BONUS
+import lotto.util.Constants.MATCH_FOUR
+import lotto.util.Constants.MATCH_SIX
+import lotto.util.Constants.MATCH_THREE
+import lotto.util.Constants.PERCENTAGE_MULTIPLIER
 import java.math.BigDecimal
 import java.math.RoundingMode
 
-private const val THREE_MATCH_PRIZE = 5000
-private const val FOUR_MATCH_PRIZE = 50000
-private const val FIVE_MATCH_PRIZE = 1500000
-private const val FIVE_BONUS_MATCH_PRIZE = 30000000
-private const val SIX_MATCH_PRIZE = 2000000000
-
 class LottoResultAnalyzer {
-    val analyzedLottoResults:List<Int>
+    val analyzedLottoResults: List<Int>
         get() = _analyzedLottoResults
 
     private var _analyzedLottoResults = MutableList(5) { 0 }
@@ -27,11 +27,11 @@ class LottoResultAnalyzer {
 
     private fun analyzeNumber(number: Int) {
         when (number) {
-            3 -> _analyzedLottoResults[0]++
-            4 -> _analyzedLottoResults[1]++
-            5 -> _analyzedLottoResults[2]++
-            100 -> _analyzedLottoResults[3]++
-            6 -> _analyzedLottoResults[4]++
+            MATCH_THREE -> _analyzedLottoResults[0]++
+            MATCH_FOUR -> _analyzedLottoResults[1]++
+            MATCH_FIVE -> _analyzedLottoResults[2]++
+            MATCH_FIVE_BONUS -> _analyzedLottoResults[3]++
+            MATCH_SIX -> _analyzedLottoResults[4]++
         }
     }
 
@@ -45,7 +45,16 @@ class LottoResultAnalyzer {
     fun calculateProfitRate(purchaseAmount: Int) {
         sumProfit()
         val divResult = _profit.toDouble().div(purchaseAmount)
-        val rate = BigDecimal(divResult).setScale(2, RoundingMode.HALF_DOWN).toDouble() * 100
+        val rate = BigDecimal(divResult).setScale(2, RoundingMode.HALF_DOWN).toDouble() * PERCENTAGE_MULTIPLIER
         _profitRate = rate
+    }
+
+    companion object {
+        private const val THREE_MATCH_PRIZE = 5000
+        private const val FOUR_MATCH_PRIZE = 50000
+        private const val FIVE_MATCH_PRIZE = 1500000
+        private const val FIVE_BONUS_MATCH_PRIZE = 30000000
+        private const val SIX_MATCH_PRIZE = 2000000000
+
     }
 }
