+++ src/main/kotlin/lotto/model/LottoTicketCount.kt
@@ -1,6 +1,6 @@
 package lotto.model
 
-import lotto.controller.LottoProgram.Companion.LOTTO_PURCHASE_COST
+import lotto.util.Constants.LOTTO_PURCHASE_COST
 
 class LottoTicketCount(private val _purchase: String) {
     val lottoCount: Int
@@ -9,7 +9,4 @@ class LottoTicketCount(private val _purchase: String) {
     private fun getLottoCount(purchase: Int):Int {
        return  purchase / LOTTO_PURCHASE_COST
     }
-
-
-
 }
