+++ src/main/kotlin/lotto/util/Constants.kt
@@ -2,4 +2,18 @@ package lotto.util
 
 object Constants {
     const val LOTTO_SIZE = 6
+    const val LOTTO_PURCHASE_COST = 1000
+    const val ZERO_NUM = 0
+    const val LOTTO_RANGE_START = 1
+    const val LOTTO_RANGE_END = 45
+    const val PERCENTAGE_MULTIPLIER = 100
+    const val DELIMITER = ','
+
+    const val MATCH_THREE = 3
+    const val MATCH_FOUR = 4
+    const val MATCH_FIVE = 5
+    const val MATCH_FIVE_BONUS = 100
+    const val MATCH_SIX = 6
+
+
 }
